# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Registry API
  description: >
    Centralized registry for MCP servers and AI agents.
    Exposes a versioned HTTP API; all UIs are clients of this API.
  version: 0.4.0-rc0
  contact:
    name: AI Registry Team
servers:
  - url: /
    description: Current host
...
paths:
...
  /api/v1/grants:
    get:
      operationId: listGlobalGrants
      summary: List global (all-publishers) role grants (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: Global grants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleGrantList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createGlobalGrant
      summary: Create a global role grant (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantRequest" }
      responses:
        "201":
          description: The created grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleGrant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/grants/{id}:
    delete:
      operationId: deleteGlobalGrant
      summary: Delete a global role grant (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/publishers/{slug}/grants:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listPublisherGrants
      summary: List a publisher's role grants (publisher Admin or Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: The publisher's grants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleGrantList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createPublisherGrant
      summary: Grant a role on a publisher (publisher Admin or Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GrantRequest" }
      responses:
        "201":
          description: The created grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleGrant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/publishers/{slug}/grants/{id}:
    delete:
      operationId: deletePublisherGrant
      summary: Revoke a role grant on a publisher (publisher Admin or Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Revoked
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Publishers ────────────────────────────────────────────────────────────

  # ── Review queue ───────────────────────────────────────────────────────────
...
  /api/v1/publishers/{slug}:
    get:
      operationId: getPublisher
      summary: Get a publisher by slug
      tags: [publishers]
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Publisher detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchPublisher
      summary: Update a publisher's metadata (publisher Admin)
      description: >
        Updates the publisher's display name and contact. Requires the Admin
        role on this publisher, or global Server Admin. Returns 403 for any
        lesser role, 401 when unauthenticated.
      tags: [publishers]
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display name for the publisher
                contact:
                  type: string
                  description: Contact email or URL
      responses:
        "200":
          description: Updated publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deletePublisher
      summary: Delete a publisher (admin only)
      description: Hard-deletes the publisher. Returns 409 if the publisher still owns active MCP servers or agents.
      tags: [publishers]
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "204":
          description: Publisher deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ai_registry_session
      description: >
        Registry session cookie set by the login endpoints. Secure + HttpOnly;
        the browser sends it automatically. There is no bearer token.
  parameters:
    SlugParam:
      name: slug
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    RoleGrant:
      type: object
      required: [id, principal_type, principal_id, role, source, created_at]
      properties:
        id: { type: string }
        principal_type: { type: string, enum: [user, group] }
        principal_id: { type: string }
        principal_label:
          type: string
          description: Human handle for the principal — a user's email or a group's slug.
        publisher_id:
          type: string
          description: Empty for a global (all-publishers) grant.
        role: { type: string, enum: [viewer, editor, reviewer, admin] }
        source: { type: string, enum: [api, config] }
        created_at: { type: string, format: date-time }
    RoleGrantList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/RoleGrant" }
    GrantRequest:
      type: object
      required: [principal_type, principal_id, role]
      properties:
        principal_type: { type: string, enum: [user, group] }
        principal_id: { type: string }
        role: { type: string, enum: [viewer, editor, reviewer, admin] }
...
    Publisher:
      type: object
      required: [id, slug, name, verified, created_at, updated_at]
      properties:
        id:
          type: string
          description: ULID
        slug:
          type: string
        name:
          type: string
        contact:
          type: string
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but lacks required scope
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Duplicate slug or version
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request body failed schema validation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...